+++ harvey/__init__.py
@@ -4,6 +4,5 @@ from harvey.git import Git
 from harvey.globals import Global
 from harvey.messages import Message
 from harvey.pipelines import Pipeline
-from harvey.stages import Deploy
 from harvey.utils import Utils
 from harvey.webhooks import Webhook